Create and join subprojects (backend)
When creating a project, the user can choose to make it a subproject of an existing project. They can also specify whether or not all members of the parent project should be automatically added to the subproject. When a user joins a subproject, they are automatically added to all ancestor projects.